requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
openai==1.30.1
langchain-community==0.0.38
pypdf==4.2.0
httpx==0.27.0